requirements.txt

yt-dlp
zhconv
pyyaml

# AI
google-api-python-client
google-cloud-aiplatform >= 1.38
anthropic

mlx-whisper